Implement max_batch_size method for Sealevel

+++ rust/main/submitter/src/chain_tx_adapter/adapter.rs
@@ -42,7 +42,7 @@ pub trait AdaptsChain: Send + Sync {
 
     /// Returns the maximum batch size for this chain. Used to decide how many payloads to batch together, as well as
     /// how many network calls to perform in parallel
-    fn max_batch_size(&self) -> usize;
+    fn max_batch_size(&self) -> u32;
 
     // methods below are excluded from the MVP
